+++ app/src/main/java/com/hongmyeoun/goldcalc/navigation/NavControl.kt
@@ -3,8 +3,6 @@ package com.hongmyeoun.goldcalc.navigation
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.animation.Crossfade
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -12,7 +10,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
@@ -20,8 +17,7 @@ import com.hongmyeoun.goldcalc.model.roomDB.character.CharacterRepository
 import com.hongmyeoun.goldcalc.view.characterDetail.CharacterDetailUI
 import com.hongmyeoun.goldcalc.view.common.LoadingScreen
 import com.hongmyeoun.goldcalc.view.goldCheck.setting.GoldSetting
-import com.hongmyeoun.goldcalc.view.main.MainScreen
-import com.hongmyeoun.goldcalc.view.main.characterCard.CharacterCard
+import com.hongmyeoun.goldcalc.view.home.HomeView
 import com.hongmyeoun.goldcalc.view.search.SearchUI
 import com.hongmyeoun.goldcalc.view.setting.SettingUI
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.AbyssDungeonVM
@@ -29,8 +25,6 @@ import com.hongmyeoun.goldcalc.viewModel.goldCheck.CommandBossVM
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.EpicRaidVM
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.GoldSettingVM
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.KazerothRaidVM
-import com.hongmyeoun.goldcalc.viewModel.main.CharacterCardVM
-import com.hongmyeoun.goldcalc.viewModel.main.CharacterListVM
 import kotlinx.coroutines.delay
 
 @RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
@@ -40,32 +34,13 @@ fun NavControl(characterRepository: CharacterRepository) {
 
     NavHost(
         navController = navController,
-        startDestination = Screen.Main.route,
+        startDestination = Screen.Home.route,
     ) {
-        composable(Screen.Main.route) {
-            val characterListVM: CharacterListVM = hiltViewModel()
-
-            MainScreen(
-                characterListVM = characterListVM,
-                navController = navController
-            ) { modifier ->
-                val characterList by characterListVM.characters.collectAsState()
-                val isLoading by characterListVM.isLoading.collectAsState()
-
-                LazyColumn(modifier = modifier) {
-                    items(characterList, key = { item -> item.name }) {
-                        val characterName = it.name
-                        val characterCardVM = remember { CharacterCardVM(characterRepository, characterName) }
-
-                        CharacterCard(
-                            navController = navController,
-                            cardViewModel = characterCardVM,
-                            isLoading = isLoading
-                        )
-                    }
-                }
-
-            }
+        composable(Screen.Home.route) {
+            HomeView(
+                navController = navController,
+                characterRepository = characterRepository
+            )
         }
         composable(Screen.Homework.route) {
             val charName = it.arguments?.getString("charName") ?: "ERROR"
