+++ app/src/main/java/com/hongmyeoun/goldcalc/navigation/Screen.kt
@@ -1,7 +1,7 @@
 package com.hongmyeoun.goldcalc.navigation
 
 sealed class Screen(val route: String) {
-    object Main : Screen("Main")
+    object Home : Screen("Main")
 
     object Homework: Screen("Homework/{charName}") {
         fun moveWithName(charName: String) = "Homework/$charName"
